5-15-15
*In This Commit....

**Responsiveness
 I have done away with trying to make the site responsive for all screen sizes because of the ramifications to the image insertion that Jack is trying to facilitate. I have maintained some responsiveness, but only down to about 80 percent of the widest size-from there it locks out and the window encroaches.

**Removed Extra Style Sheet
I removed the haf-base styles.css from the project, because of the change in direction from the responsiveness goals.

**Forms
Added in the form fields (Jack or Brad, please check these form fields--I couldn't find Jack's AddItemForm.ejs on Git, so I watched a couple how to's and lifted code from the internet to construct.

**Upper Nav and Logo
Added in upper nav and logo at very top. I may still restyle these if I have time.

**Mike Still Doing 1
Making a the thumbnail version of the hat description. This is the div that will be a thinner, summarized version of the hat, that will show only a small picture of the hat.

**Mike Still Doing 2
Making a spring sale link on home page as well a new section that will tell about the spring sale.