2105.05.20.11:17
haf-large-style.css
Removed all but two blocks of code for the detail bar, one block for the model photo on the left and one for the model photo on the right. Removed the ordering from the names of the remaining blocks. Removed URL paths for background images for picture containers.

functions.js
Added code to poke URLs for the current background images into the CSS for the left or right detail bars. Removed debugging statements.