2015.05.21.16:25  Commited files for updated design of thumbnails.
functions.js
Removed debugging information.

thumbnail.ejs
Updated structure to current design. Changed array key references from capital to lower case letters.

haf-thumbnl-styles.css
Mike added an outer container "thumbnail-1" for the thumbnail and set a height for it. The height constrained the height when the detail was swapped in. Removed the height for "thumbnail-1" and "thumbnail-2."

haf-thumbnail-styles.css
Mike added a container thumbnail-1 for