2015.06.09.13:38 - Added functionality to display the number of hats in the shopping cart.
cart.js
Added code to get the number of hats in the shopping cart and change the display when items are deleted and when the cart is emptied after the order is written to the database. Added code to empty the shopping cart.

functions.js
Added code to get and display the number of hats in the shopping cart and change the display when items are added to the shopping cart. Added code to open and close shopping cart page.

start.html
Added code to open and close shopping cart page.